# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reeve",
    "description": "Fleet desired-state manager (Margo-inspired). Human API consumed by the reeve UI, plus the device-facing wire surface.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/durability/status": {
      "get": {
        "tags": [
          "durability"
        ],
        "summary": "GET /api/durability/status — the §9.4 API surface (viewer+; the UI\nrenders \"last verified restore\" and the degraded flag from this).",
        "operationId": "durability_status",
        "responses": {
          "200": {
            "description": "Durability tier status: degraded flag, upload lag, last verified restore",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurabilityStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below viewer role"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DurabilityStatus": {
        "type": "object",
        "description": "The queryable status shape (§9.2 degraded flag, §9.3 upload lag,\n§9.4 \"last verified restore\"). A deployment whose verify-restore\nhas never succeeded reads as having NO durability tier (§9.4) —\n`effective_tier` encodes exactly that rule for the UI.",
        "required": [
          "tier",
          "degraded",
          "pending_changesets"
        ],
        "properties": {
          "degraded": {
            "type": "boolean"
          },
          "generation": {
            "type": [
              "string",
              "null"
            ],
            "description": "Current (last successfully shipped) snapshot generation."
          },
          "last_changeset_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "last_changeset_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_snapshot_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "last_verify": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VerifySummary"
              }
            ]
          },
          "pending_changesets": {
            "type": "integer",
            "minimum": 0
          },
          "snapshot_age_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tier": {
            "type": "string"
          }
        }
      },
      "DurabilityStatusResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DurabilityStatus"
          },
          {
            "type": "object",
            "required": [
              "effective_tier",
              "epoch"
            ],
            "properties": {
              "effective_tier": {
                "type": "string",
                "description": "`none` | `snapshot` | `changeset` | `none (unverified)`."
              },
              "epoch": {
                "type": "integer",
                "format": "int32",
                "description": "Current server epoch (`settings.server_epoch`, spec/reeve/07-durability.md\n§9.5): the high 16 bits of every manifestVersion. Restore fencing\nincrements it — the ops UI surfaces it next to durability posture.",
                "minimum": 0
              }
            }
          }
        ],
        "description": "`GET /api/durability/status` body: [`DurabilityStatus`] plus the\n§9.4 `effective_tier` (an unverified tier reads as none)."
      },
...
      "VerifySummary": {
        "type": "object",
        "description": "Latest recorded run, for the status surface (\"last verified\nrestore: <when>\", §9.4).",
        "required": [
          "finished_at",
          "outcome"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "type": "integer",
            "format": "int64"
          },
          "generation": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "outcome": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}